requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.35
python-multipart>=0.0.20
jinja2>=3.1.5
passlib>=1.7.4
email-validator>=2.2.0
python-dotenv>=1.0.1
bcrypt>=4.1.2
python-jose>=3.3.0
itsdangerous>=2.2.0
starlette>=0.35.0
requests>=2.31.0
PyJWT>=2.8.0